Update README.md
Instacart is an innovative online grocery store that has been operating through an app. To enhance their business's growth and improve their sales patterns, they are looking for an in-depth analysis of their data. They require an exploratory analysis of their data to derive insights and form strategies for better customer segmentation. Instacart stakeholders are keen to understand the variety of customers in their database and their purchasing behaviours. With this information, Instacart aims to launch a targeted marketing strategy that can meet the customer's needs and satisfaction.
Key Questions
What are the busiest days of the week and hours of the day?
Are there particular times of the day when people spend the most money?
Are there simpler price range groupings to help direct marketing efforts?
Are there certain types of products that are more popular than others?
How do the ordering behaviours of different types of customers differ?
What’s the distribution among users regarding
 their brand loyalty?
Are there differences in ordering habits based on a customer’s loyalty status?
Are there differences in ordering habits based on a customer’s region?
Is there a connection between age and family status in terms of ordering habits?
What different classifications does the demographic information suggest?
What differences can you find in the ordering habits of different customer profiles?
Data
(Data files not uploaded to GitHub due to size.)

It is worth noting that Instacart is an actual company that has made their data publicly available. However, for the purposes of this project, the data sets provided were specifically created for this project by Career Foundry. The fabricated data sets included various data points that were relevant to the scope of this project.
Customers
Departments
Orders
Orders_products_prior
Products
Tools
The following Python libraries and other tools were used in this project:

Jupyter Notebook: for writing and executing Python code
Pandas: for data analization
Seaborn: for data visualization
Matplotlib: for data visualization
Numpy: for mathematical equations
SciPy: for mathematical equations
ChatGPT: for assisting in creating code that allowed me to normalize data and format charts in the ways I wanted them to appear.
Folders
The project files are divided between the following folders:

Project Management: Project Brief
Data (Data files not uploaded to GitHub due to size.)
Scripts: The Jupyter notebooks containing the coding for the analysis.
Analysis: The visualizations used for developing and explaining insights.
Sent to Client: The final report presented in Excel.